Added a homing missile and way to shoot them
To shoot the missiles it will check if any enemies are within the vicinity/range. After it will spawn some missiles and will track/follow it's target. Will play an explosion effect on impact.

If there is not a target in range the missile will fire in a random direction and explode after some time.